goiymonan-backend/rl/q-learning/requirements.txt
fastapi==0.115.0
uvicorn==0.30.1
pydantic==2.9.2
psycopg2-binary
